chore: extract useRefCallback
